// src/app/data/model/Edamam.ts
export interface Ingredient {
    text: string;
    weight: number;
}
// ...
export interface TotalNutrient {
    eNERC_KCAL: ENERC_KCAL;
    fAT: FAT;
    fASAT: FASAT;
    fATRN: FATRN;
    fAMS: FAM;
    fAPU: FAPU;
    cHOCDF: CHOCDF;
    fIBTG: FIBTG;
    sUGAR: SUGAR;
    pROCNT: PROCNT;
    cHOLE: CHOLE;
    nA: NA;
    cA: CA;
    mG: MG;
    k: K;
    fE: FE;
    zN: ZN;
    p: P;
    vITA_RAE: VITA_RAE;
    vITC: VITC;
    tHIA: THIA;
    rIBF: RIBF;
    nIA: NIA;
    vITB6A: VITB6A;
    FOLDfE: FOLDFE;
    fOLFD: FOLFD;
    fOLAC: FOLAC;
    vITB12: VITB12;
    vITD: VITD;
    tOCPHA: TOCPHA;
    vITK1: VITK1;
    wATER: WATER;
}
// ...
export interface TotalDaily {
    eNERC_KCAL: ENERC_KCAL;
    fAT: FAT;
    fASAT: FASAT;
    cHOCDF: CHOCDF;
    fIBTG: FIBTG;
    pROCNT: PROCNT;
    cHOLE: CHOLE;
    nA: NA;
    cA: CA;
    mG: MG;
    k: K;
    fE: FE;
    zN: ZN;
    p: P;
    vITA_RAE: VITA_RAE;
    vITC: VITC;
    tHIA: THIA;
    rIBF: RIBF;
    nIA: NIA;
    vITB6A: VITB6A;
    FOLDfE: FOLDFE;
    vITB12: VITB12;
    vITD: VITD;
    tOCPHA: TOCPHA;
    vITK1: VITK1;
}

export interface Sub {
    label: string;
    tag: string;
    schemaOrgTag: string;
    total: number;
    hasRDI: boolean;
    daily: number;
    unit: string;
}

export interface Digest {
    label: string;
    tag: string;
    schemaOrgTag: string;
    total: number;
    hasRDI: boolean;
    daily: number;
    unit: string;
    sub: Sub[];
}

export interface Recipe {
    uri: string;
    label: string;
    image: string;
    source: string;
    url: string;
    shareAs: string;
    yield: number;
    dietLabels: string[];
    healthLabels: string[];
    cautions: string[];
    ingredientLines: string[];
    ingredients: Ingredient[];
    calories: number;
    totalWeight: number;
    totalTime: number;
    totalNutrients: TotalNutrient;
    totalDaily: TotalDaily;
    digest: Digest[];
}

export interface Hit {
    recipe: Recipe;
    bookmarked: boolean;
    bought: boolean;
}